Revert "refactor: move topAppBar to components"
This reverts commit 92fdc919c0ecf7222bb482d5e2d55aa2bd1cfa86.
